pyproject.toml
[tool.poetry]
name = "song-metadata-checker"
version = "0.1.0"
description = ""
authors = ["yicone <yicone@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
python-dotenv = "^1.2.0"
pydantic = "^2.12.3"

[tool.poetry.group.dev.dependencies]
# Development and testing dependencies
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.12.0"
ruff = "^0.8.0"
mypy = "^1.10.1"
pyyaml = "^6.0.3"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
